Changed endpoint to new cluster, added QueryResult object to format query data
