+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 server.port=8082
-opensearch_host=${OPENSEARCH_HOST:a4059048841004a80a3e273b13fe4b57-228758828.us-west-2.elb.amazonaws.com}
+opensearch_host=${OPENSEARCH_HOST:ac4038229ae5842ffa50c08c2ebab0be-623641807.us-west-2.elb.amazonaws.com}
 opensearch_port=${OPENSEARCH_PORT:9200}
